main: fix max-io-requests spelling in warning text
An incorrect spelling of max-io-requests was used in creating the
warning message text. Due to conversion to unsigned, the code would
crash due to bad_lexical_cast exception. The spelling of this
configuration name was fixed in the past (44f3ad836bf), but only in the
'if' condition.

Fix by using the correct spelling.

Closes #8963