scylla_io_setup: check root privilege on root mode
This is side effect of allowing to run scylla_io_setup in nonroot mode, the script able to run in non-root user even the installation is not nonroot mode.
Result of that, the script finally failed to write io_properties.yaml and causes permission denied.
Since the evaluation takes long time, we should run permission check before starting it.

We need to add root privilege check again, but skip it on nonroot mode.

Fixes #8915