sstables: use read-aheads in the index reader
Currently, when advancing one of index_reader's bounds,
we're creating a new index_consume_entry_context with a new
underlying file input_stream for each new page.

For either bound, the streams can be reused, because
the indexes of pages that we are reading are never
decreasing.

This patch adds a index_consume_entry_context to each of
index_reader's bounds, so that for each new page, the same
file input_stream is used.
As a result, when reading consecutive pages, the reads that
follow the first one can be satisfied by the input_stream's
read aheads, decreasing the number of blocking reads and
increasing the throughput of the index_reader.

Fixes #2388

Signed-off-by: Wojciech Mitros <wojciech.mitros@scylladb.com>