sstables: Do not populate page cache when searching in promoted index for "bypass cache" reads
Reads which bypass cache will use a private temporary instance of cached_file
which dies together with the index cursor.

The cursor still needs a cached_file with cachig layer. Binary searching needs
caching for performance, some of the pages will be reused. Another reason to still
use cached_file is to work with a common interface, and reusing it requires minimal changes.