scylla_util.py - Azure Instance Checks
This patch replaces the Azure IMDS endpoint to 169.254.169.254, which is the only correct value to be used as in Azure documentation.
It replaces the socket.getaddrinfo() call to curl(), as getaddrinfo() may potentially raise false-positives with GCP instances.

Finally, it corrects a contradiction (i.e: if False and True) when checking whether the instance is supported.