test/alternator: test another corner case of TTL
Usually the TTL feature's expiration-time attribute is a schema-less
attribute, implemented in Alternator as a JSON-serialized item in a
bigger map column. However, key attributes are a special case because
they are implemented as separate columns. We already had test cases
showing that this case works too - for the case of hash and range keys.

In this test we test another possibility of an attribute that is
implemented as a schema column - the case of an LSI key.

As the other TTL tests, this test too passes on DynamoDB but xfails on
Alternator because the TTL feature is not yet implemented.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>